src/dialog: rename simple_row to simple_palette_row
