+++ src/dialog/add_new_palette_dialog.rs
@@ -12,7 +12,7 @@ use crate::i18n::{i18n, i18n_f};
 use crate::model::color::Color;
 use crate::pages::color_square::ColorSquare;
 
-use super::simple_row::SimplePaletteRow;
+use super::simple_palette_row::SimplePaletteRow;
 
 mod imp {
     use super::*;
