+++ src/dialog/simple_palette_row.rs
